+++ src/index.ts
@@ -17,7 +17,10 @@ export type {
   MobileImageMounterService,
   NotificationProxyService,
   MobileConfigService,
+  PowerAssertionService,
+  PowerAssertionOptions,
   SpringboardService,
+  WebInspectorService,
   SyslogService,
   SocketInfo,
   TunnelResult,
@@ -27,8 +30,11 @@ export type {
   MobileImageMounterServiceWithConnection,
   NotificationProxyServiceWithConnection,
   MobileConfigServiceWithConnection,
+  PowerAssertionServiceWithConnection,
   SpringboardServiceWithConnection,
+  WebInspectorServiceWithConnection,
 } from './lib/types.js';
+export { PowerAssertionType } from './lib/types.js';
 export {
   createUsbmux,
   Services,